mesh_on_cloud_run/src/llm_proxy/requirements.txt
google-genai
google-generativeai
pydantic
uvicorn[standard]
gunicorn
fastapi
